multi router+graph: Realtime prune disabled channels.
Prior to this commit the pruneZombieChannels logic was to prune
both channels with expired updates and channels which their both
ChannelEdgePolicy have their disabled bit on (When
AssumeChannelValid=false).

This commit leverages the fact that we can identify the later case in
realtime (when the  ChannelEdgePolicy update lands) and moves the prune
logic of disabled channels to the point where the update is
received.

The advantages of this change are:
1. Graph is more up to date.
2. We can optimize the pruneZombieChannels function that is very
expensive right now as it iterates all the graph on startup.
The optimization is now possible since we don't have to go over all
the graph for pruning expired channels and we can use
ChanUpdatesInHorizon function instead.