routing+channeldb: Record failed routes.
In this commit a new field "FailedRoutes" is added as a persistent field
on Payment structure.
This field contains the routes of previous failed payment attempts and
is populated when the payment is failed.
The purpose of it is to be able to analyze payment failures that caused
by routing problems and get further insights about the user's graph.