Add the ability to persist compact filters to disks
When synchronizing using neutrino, if the downloaded filters are not stored
persistently and lnd is stopped before it finishes, lnd will have to
download again all the needed filters.
This commit adds the boolean configuration flag PersistToDisk. If set to
true, neutrino will save to disk the filters as it download them.