Let SettleHodlInvoice succeed even if the invoice is open.
SettleHoldInvoice is called after the payee knows that she received htlcs.
There are cases when the payee accept to accept an amount lower that the
invoice's amount - for instance if she wants to make a discount or if the
payee wants to pay a part of the amount to a node which just created a new
channel for her. Generally the hold invoice is used by a payee who wants
to bypass the automatic mechanisms of lnd. So there is no reason to stop
her to settle the invoice when the sum of the htlcs is less that the
invoice's amount.