fundingmanager: skip channel confirmation.
This commit exposes a function to skip the funding confirmation of
a pending channel. Specifically it advances the funding workflow and
make the channel fully operational by sending and receiving
funding locked message.When using this function one should provide a
fake short channel id which is validated to fall into a cretain
acceptance range that does not conflict with confirmed channels.