rpc: Add SkipFundingConfirmation endpoint.
This adds the rpc layer of skipping channel confirmation
operation that was introduced prior to this commit.
When both parties agree on a fake short channel id, they
can use it as a temporary id for the funding locked message
untill channel is confirmed.
This will make the channel fully operational but requires trust
between the parties.