fundingmanager: Skip fetch channel outpoint.
For local channel announcement we can skip fetching the outpoint
as we already have it and we don't need to validate it.
Although the gossiper has the information to distinguish between
remote and local announcements this flag is not propagated to the
router. Instead a check is made if we are a side in the channel and
a proof was't provided. This means it is our announcement that probably
was sent during funding.