fundingmanager: Allow skip confirmation.
This commit exposes a configuration flag for the channel receiver
to skip the funding confirmation of a pending channel.
Specifically it advances the funding workflow and makes the channel
fully operational by sending and receiving funding locked message.
Since there is no risk at the initiator side, when receiving zero amount
of required confirmation from the reciever it advances the workflow
as well.
During the time where the channel is operational but not confirmed both
parties use a convention where the first 58 bits of the random pending
id serves as the short channel id as long as it does not conflict with
existing ids. This convention ensures that the block height of the
generated id is less than 50000 and doesn't require any extra RPC
layer.