channeldb+lnwallet: delete owed remote updates when signing
In the same vein, this commit also modifies the state restore
logic to not insert the corresponding update_add_htlc if it exists
on the remoteCommitment, but not on the pendingRemoteCommitment.
This would mean that we've signed an update cancelling it.

Specifically, this fixes the following state transition:

Alice                   Bob
       <----add-------
       <----sig-------
       -----rev------>
       -----sig------>
       <----rev-------
* The HTLC is locked in on both sides here *

       ----settle---->
       -----sig------>
       <----rev-------
* Bob stores the settle here, he must sign it *

       <----sig-------
* Bob signs the settle, but it isn't deleted *

       -----add------>
       -----sig------>
       <----rev-------
* Bob will delete the stored settle as his revocation wipes all
stored updates *

* Bob restarts *

       -----rev------>
* Alice opens up the revocation window *

       <----sig-------
* Bob's commitment signature is invalid because the state restore
logic assumes it can find the add's corresponding settle, but it can't.
Bob will fail to sign a commitment here. *